+++ packages/driver/cypress/e2e/e2e/origin/basic_login.cy.ts
@@ -104,9 +104,7 @@ describe('basic login', { browser: '!webkit' }, () => {
         }, {
           validate: () => {
             cy.window().then((win) => {
-              const cypressAuthToken = win.sessionStorage.getItem('cypressAuthToken')
-
-              return !!cypressAuthToken
+              expect(win.sessionStorage.getItem('cypressAuthToken')).to.exist
             })
           },
         })