+++ packages/errors/test/unit/visualSnapshotErrors_spec.ts
@@ -1001,6 +1001,11 @@ describe('visual error templates', () => {
         default: [],
       }
     },
+    EXPERIMENTAL_SESSION_AND_ORIGIN_REMOVED: () => {
+      return {
+        default: [],
+      }
+    },
     EXPERIMENTAL_SHADOW_DOM_REMOVED: () => {
       return {
         default: [],
@@ -1212,9 +1217,21 @@ describe('visual error templates', () => {
       }
     },
 
+    EXPERIMENTAL_RUN_ALL_SPECS_E2E_ONLY: () => {
+      return {
+        default: [],
+      }
+    },
+
     BROWSER_UNSUPPORTED_LAUNCH_OPTION: () => {
       return {
         default: ['electron', ['env']],
+      }
+    },
+
+    EXPERIMENTAL_ORIGIN_DEPENDENCIES_E2E_ONLY: () => {
+      return {
+        default: [],
       }
     },
   })