+++ system-tests/projects/run-all-specs/cypress-invalid-root.config.js
@@ -1,0 +1,8 @@
+const { defineConfig } = require('cypress')
+
+module.exports = defineConfig({
+  experimentalRunAllSpecs: true,
+  e2e: {
+    supportFile: false,
+  },
+})